Refactor host system of local DNS server
Move the rules parser to the V2Ray. Considering we parse the rules in
the strmatcher package, programers only need to care about the rules
themselves now.

In the previous version, we prase the configure file in the V2Ctl and
only send the base rules to the V2Ray. It works, but is hard
to add complex rules and to reuse code.